requirements.txt
requests==2.31.0
openai==1.12.0
python-dotenv==1.0.0
tenacity==8.2.3
ratelimit==2.2.1
sqlalchemy==2.0.27
pandas==2.1.4
tqdm==4.66.1
black==23.12.1 
openpyxl==3.1.5
yfinance==0.2.37